+++ Changes.md
@@ -22,6 +22,7 @@ Fixes
   - Fixed failure to change sampling pattern per frame.
   - Fixed Resolution Multiplier support.
   - Fixed UI visibility of `angle` parameter for distant lights, which was previously hidden.
+  - Fixed export of `Varying` primitive variables on meshes (#5781).
 
 1.4.0.0b6 (relative to 1.4.0.0b5)
 =========